Fix up a large number of warnings from GCC4.2, mostly relating to conversion from const char * to char *.
This fixup has been executed carefully, in all cases by checking what the
code is actually doing.  Where appropriate, function prototypes were changed
from char * to const char *; there was a small number of cases where string
constants were typecast to (char *), but only where it was clear that the
code did not attempt to modify the string and that alternative of rewriting
the code was worse.  In two situations, code was rewritten where it was
typecasting was clearly inappropriate: the symID constructor in
rpsl/rpsl/symbols.hh was changed in this way; also a new class member was
added to the RtConfig class (and its children) called
::returnPermitOrDeny().  Possibly this is slight overkill, but it works and
does the right thing while creating code efficiency.

The code now compiles with no warnings on GCC4.2 with the default warning
settings.