restrict() in NormalExpression seems not to have been thought out well. It didn't work properly in anything but simple v4/v6 prefixes, probably needs some serious rework.
This time, the bug is that it set the singleton_flag to PRFX/IPV6_PRFX
if it found a non-empty prefix list inside the term. It didn't account,
however, for terms containing a prefix list *and* something else (e.g.
a community) or for expressions containing multiple, different, terms.

Stop being that smart and just copy the singleton_flag from the
previous, non-restricted(?) normal expression.

Fix provided by Faidon Liambotis <paravoid@debian.org>