Fix for ticket:12, ticket:14 and a reversion of some of the changes from ticket:35.
Reported by S.P.Zeidler on the mailing list:

Hagen Boehms cache patch breaks RAWhoisClient::getSet for all objects but
RSSets and ASSets which are also getting served by this function (so that
eg "peval -compressed fltr-bogons" no longer works.

spz's patch had a minor problem in that it only worked properly for as-sets
and rs-sets.  Revert one of the changes from r216 to fix this (ticket:14)

Thanks to spz for original patch and davidc for noting regression.