Do not use -V on RIPE/Bird WHOIS queries
irtoolset currently sends -V <ProjectVersion> for all of its WHOIS
queries.

RIPE's Database Query Reference Manual, section "Access Control for
Queries" mentions that this is used by proxies to indicate the client's
IP address in the syntax
    -V <version>,<ipv4-address>
which is obviously not something that irrtoolset does or needs.

The help output in RIPE's own WHOIS server, however, mentions:
    -V<client-tag>   Sends information about the client to the server.
which is probably the reason for this misconception.

(in HTTP terms, that would be a Via vs. User-Agent header)

The WHOIS server's source indicates that the former is the proper
meaning of the flag and even irrtoolset clients that try to pass that
without being explicitelly marked as trusted proxies.

Hence, stop sending the flag altogether.