Add travis CI testing file for easier fault finding
Travis CI is a valuable tool to validate pull requests and gain a better
understanding of the project's health. Not everybody will attempt
building source code they contribute, or have a standardized build
environment. Enabling Travis on this project will provide coherency in
that regard and provide semi-instant feedback on pull requests.