Fix for https://github.com/irrtoolset/irrtoolset/issues/25 (#38)
see the discussion in #25:

this patch works around a more general problem described #26, namely that the code is too naive to understand expressions involving prefix lists and other types of terms. This patch doesn't fix the underlying problem: it just removes a sanity check which causes some legitimate expressions to break.